requirements.txt
asgiref==3.9.1
channels==4.2.2
channels_redis==4.3.0
click==8.2.1
Django==5.2.4
django-dirtyfields==1.9.7
django-filter==25.1
django-model-utils==5.0.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
h11==0.16.0
msgpack==1.1.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.2.0
sqlparse==0.5.3
uvicorn==0.35.0
websocket-client==1.8.0
websockets==15.0.1
